requirements.txt
# Core AI/ML Libraries
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.0.2
autogen>=0.9.0
crewai>=0.11.0

# OpenAI and Anthropic
openai>=1.10.0
anthropic>=0.7.0

# Vector Database
pinecone-client>=2.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
websockets>=11.0

# Database
pymongo>=4.5.0
redis>=5.0.0

# Message Queue
kafka-python>=2.0.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring
prometheus-client>=0.17.0

# Security
cryptography>=41.0.0

# Development
jupyter>=1.0.0
ipython>=8.0.0

# Language Models
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.20.0

# Vector Databases
chromadb>=0.4.0
weaviate-client>=3.25.0

# ML/DL Frameworks
tensorflow>=2.15.0
pytorch-lightning>=2.1.0
xgboost>=2.0.0

# Computer Vision
clip-interrogator>=0.6.0
diffusers>=0.24.0

# NLP
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.0

# Blockchain/Web3
web3>=6.0.0
eth-account>=0.9.0
py-solc-x>=2.0.0

# API/Server
celery>=5.3.0
kombu>=5.3.0

# Monitoring/Logging
wandb>=0.16.0
mlflow>=2.8.0
tensorboard>=2.15.0

# Testing
hypothesis>=6.90.0

# Development
pre-commit>=3.5.0 